First draft of final project
"First draft" of my final project due on 2/18/18. I started on it pretty much right after we brainstormed what it would be. I've done a bit of javascript for it, especially in the last few days. Most of the css and html is done. Most of the pages still need content though. I understand that my website has just a ton of content text that I will have to input. Some of it, I can get online and the rest, I have to copy from my physical book. I understand that will take a huge amount of time. The README.txt says which pages currently have content and which don't. The README.txt also says what else I'm planning to do, and a stretch goal.